+++ NEWS.md
@@ -2,6 +2,7 @@
 
 + option of using leaflet to see experimental plots on a satellite image of the field was added to `viz()` 
 + `create_strips()` (an internal function) was overhauled to pass R CMD check on Fedora with ATLAS. Specifically, `st_tilt()`, which used to be used in the function, has been abandoned.
++ fixed the bug in `viz()` which presented `gc_rate`
 
 # ofpetrial 0.1